+++ src/FsCodec.Box/CoreCodec.fs
@@ -18,11 +18,11 @@ type Codec private () =
         (   encoder,
             // <summary>Maps from the TypeShape <c>UnionConverter</c> <c>'Contract</c> case the Event has been mapped to (with the raw event data as context)
             // to the <c>'Event</c> representation (typically a Discriminated Union) that is to be presented to the programming model.</summary>
-            up : FsCodec.ITimelineEvent<'Body> * 'Contract -> 'Event,
+            up : struct (FsCodec.ITimelineEvent<'Body> * 'Contract) -> 'Event,
             // <summary>Maps a fresh Event resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c><br/>
             // The function is also expected to derive an optional <c>meta</c> object that will be serialized with the same <c>encoder</c>,
             // and <c>eventId</c>, <c>correlationId</c>, <c>causationId</c> and an Event Creation<c>timestamp</c></summary>.
-            down : 'Context option * 'Event -> 'Contract * 'Meta option * Guid * string * string * DateTimeOffset option,
+            down : struct ('Context * 'Event) -> struct ('Contract * 'Meta voption * Guid * string * string * DateTimeOffset voption),
             // <summary>Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them.</summary>
             [<Optional; DefaultParameterValue(null)>] ?rejectNullaryCases)
         : FsCodec.IEventCodec<'Event, 'Body, 'Context> =
@@ -37,15 +37,16 @@ type Codec private () =
 
         { new FsCodec.IEventCodec<'Event, 'Body, 'Context> with
             member _.Encode(context, event) =
-                let (c, meta : 'Meta option, eventId, correlationId, causationId, timestamp : DateTimeOffset option) = down (context, event)
+                let struct (c, meta : 'Meta voption, eventId, correlationId, causationId, timestamp : DateTimeOffset voption) = down struct (context, event)
                 let enc = dataCodec.Encode c
-                let meta' = match meta with Some x -> encoder.Encode<'Meta> x | None -> Unchecked.defaultof<_>
-                EventData.Create(enc.CaseName, enc.Payload, meta', eventId, correlationId, causationId, ?timestamp = timestamp)
+                let meta' = match meta with ValueSome x -> encoder.Encode<'Meta> x | ValueNone -> Unchecked.defaultof<_>
+                let ts = match timestamp with ValueNone -> None | ValueSome v -> Some v
+                EventData.Create(enc.CaseName, enc.Payload, meta', eventId, correlationId, causationId, ?timestamp = ts)
 
             member _.TryDecode encoded =
                 match dataCodec.TryDecode { CaseName = encoded.EventType; Payload = encoded.Data } with
-                | None -> None
-                | Some contract -> up (encoded, contract) |> Some }
+                | None -> ValueNone
+                | Some contract -> up (encoded, contract) |> ValueSome }
 
     /// <summary>Generate an <c>IEventCodec</c> using the supplied <c>encoder</c>.<br/>
     /// Uses <c>up</c>, <c>down</c> and <c>mapCausation</c> functions to facilitate upconversion/downconversion and eventId/correlationId/causationId/timestamp mapping
@@ -56,22 +57,22 @@ type Codec private () =
         (   encoder,
             // <summary>Maps from the TypeShape <c>UnionConverter</c> <c>'Contract</c> case the Event has been mapped to (with the raw event data as context)
             // to the representation (typically a Discriminated Union) that is to be presented to the programming model.</summary>
-            up : FsCodec.ITimelineEvent<'Body> * 'Contract -> 'Event,
+            up : struct (FsCodec.ITimelineEvent<'Body> * 'Contract) -> 'Event,
             // <summary>Maps a fresh Event resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
             // The function is also expected to derive
             //   a <c>meta</c> object that will be serialized with the same options (if it's not <c>None</c>)
             //   and an Event Creation <c>timestamp</c>.</summary>
-            down : 'Event -> 'Contract * 'Meta option * DateTimeOffset option,
+            down : 'Event -> struct ('Contract * 'Meta voption * DateTimeOffset voption),
             // <summary>Uses the 'Context passed to the Encode call and the 'Meta emitted by <c>down</c> to a) the final metadata b) the <c>eventId</c> c) the <c>correlationId</c> and d) the <c>causationId</c></summary>
-            mapCausation : 'Context option * 'Meta option -> 'Meta option * Guid * string * string,
+            mapCausation : struct ('Context * 'Meta voption) -> struct ('Meta voption * Guid * string * string),
             // <summary>Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them.</summary>
             [<Optional; DefaultParameterValue(null)>] ?rejectNullaryCases)
         : FsCodec.IEventCodec<'Event, 'Body, 'Context> =
 
-        let down (context, union) =
-            let c, m, t = down union
-            let m', eventId, correlationId, causationId = mapCausation (context, m)
-            c, m', eventId, correlationId, causationId, t
+        let down struct (context, union) =
+            let struct (c, m, t) = down union
+            let struct (m', eventId, correlationId, causationId) = mapCausation (context, m)
+            struct (c, m', eventId, correlationId, causationId, t)
         Codec.Create(encoder, up = up, down = down, ?rejectNullaryCases = rejectNullaryCases)
 
     /// <summary>Generate an <code>IEventCodec</code> using the supplied <c>encoder</c>.<br/>
@@ -83,17 +84,17 @@ type Codec private () =
         (   encoder,
             // <summary>Maps from the TypeShape <c>UnionConverter</c> <c>'Contract</c> case the Event has been mapped to (with the raw event data as context)
             // to the representation (typically a Discriminated Union) that is to be presented to the programming model.</summary>
-            up : FsCodec.ITimelineEvent<'Body> * 'Contract -> 'Event,
+            up : struct (FsCodec.ITimelineEvent<'Body> * 'Contract) -> 'Event,
             // <summary>Maps a fresh <c>'Event</c> resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
             // The function is also expected to derive
             //   a <c>meta</c> object that will be serialized with the same options (if it's not <c>None</c>)
             //   and an Event Creation <c>timestamp</c>.</summary>
-            down : 'Event -> 'Contract * 'Meta option * DateTimeOffset option,
+            down : 'Event -> struct ('Contract * 'Meta voption * DateTimeOffset voption),
             // <summary>Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them.</summary>
             [<Optional; DefaultParameterValue(null)>] ?rejectNullaryCases)
-        : FsCodec.IEventCodec<'Event, 'Body, obj> =
+        : FsCodec.IEventCodec<'Event, 'Body, unit> =
 
-        let mapCausation (_context : obj, m : 'Meta option) = m, Guid.NewGuid(), null, null
+        let mapCausation struct ((), m : 'Meta voption) = struct (m, Guid.NewGuid(), null, null)
         Codec.Create(encoder, up = up, down = down, mapCausation = mapCausation, ?rejectNullaryCases = rejectNullaryCases)
 
     /// <summary>Generate an <code>IEventCodec</code> using the supplied <c>encoder</c>.<br/>
@@ -103,8 +104,8 @@ type Codec private () =
         (   encoder : TypeShape.UnionContract.IEncoder<'Body>,
             // <summary>Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them.</summary>
             [<Optional; DefaultParameterValue(null)>] ?rejectNullaryCases)
-        : FsCodec.IEventCodec<'Union, 'Body, obj> =
+        : FsCodec.IEventCodec<'Union, 'Body, unit> =
 
-        let up : FsCodec.ITimelineEvent<'Body> * 'Union -> 'Union = snd
-        let down (event : 'Union) = event, None, None
+        let up struct (_e : FsCodec.ITimelineEvent<'Body>, u : 'Union) : 'Union = u
+        let down (event : 'Union) = struct (event, ValueNone, ValueNone)
         Codec.Create(encoder, up = up, down = down, ?rejectNullaryCases = rejectNullaryCases)